refactor: isolate and parametrize some management CLI tests
