+++ influxdb_iox/tests/end_to_end_cases/mod.rs
@@ -2,6 +2,7 @@ mod debug_cli;
 mod delete_api;
 mod deletes;
 mod deployment_api;
+mod deployment_cli;
 mod flight_api;
 mod freeze;
 mod http;
@@ -15,6 +16,7 @@ mod persistence;
 mod read_api;
 mod read_cli;
 mod remote_api;
+mod remote_cli;
 mod router_api;
 mod run_cli;
 pub mod scenario;
